move ASTs around
