+++ compiler/core-new/src/plugins/documentation/convert.ts
@@ -2,7 +2,7 @@ import * as serialization from '@lona/serialization'
 
 import { Helpers } from '../../helpers'
 import { convertDeclaration } from '../tokens/convert'
-import { Token } from '../../types/tokens-ast'
+import { Token } from '../tokens/tokens-ast'
 import { nonNullable, assertNever } from '../../utils'
 
 let tokenNameElement = (kind: string, content: string) =>
